Fix game ready state and transport position after spawn
