+++ Library/RSBot.Core/Network/Handler/Agent/Cos/UpdateMountStateResponse.cs
@@ -42,6 +42,8 @@ namespace RSBot.Core.Network.Handler.Agent
                 if (!isMounted)
                 {
                     Game.Player.Vehicle = null;
+                    Game.Player.Transport = null;
+
                     return;
                 }
 
@@ -57,6 +59,10 @@ namespace RSBot.Core.Network.Handler.Agent
 
                 if (cosUniqueId == Game.Player.Fellow?.UniqueId)
                     Game.Player.Vehicle = Game.Player.Fellow;
+
+                var bionicPosition = Game.Player.Vehicle.Bionic.Position;
+                Game.Player.Vehicle.StopMoving(bionicPosition);
+                Game.Player.StopMoving(bionicPosition);
             }
 
             //Assertion: only player's are supported to have active vehicles. Think it's the same in the client.
