+++ Library/RSBot.Core/Objects/Cos/JobTransport.cs
@@ -1,10 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace RSBot.Core.Objects.Cos
+﻿namespace RSBot.Core.Objects.Cos
 {
     public class JobTransport : Cos
     {
